Modification to the Q  used in the calculation of alpha
initially it used the original Q (without re-normalizing considering only the specified neighbors of the PNSet). Now it starts with the original Q then filters only the neighbors specified in the PNSet and then re-distributes the probability so it adds up to 1 and this is the new probability used in the calculation of the alpha.